requirements.txt
pymodbus>=3.0.0
pyserial>=3.0
